app/requirements.txt
bson
flask
flask_cors
flask_login
pyjwt
pymongo
